+++ worldturtle/ChangeLog.md
@@ -1,5 +1,13 @@
 # Changelog for turtle-haskell
 
+## v0.2.0
+
+* Split `TurtleCommand` into `WorldCommand` and `WorldCommand` to help reduce
+  boilerplate of having to apply a turtle to a command for every stage of a
+  command block.
+* Added `runWorld`, `runTurtle`, `run`, and `(>/>)` functions.
+* `circle` command split into `arc` and `circle` commands.
+
 ## v0.1.2
 
 * Added the `branch` function.